cleaned up project structure and broke down employee pages code into multiple files
